Agent: implement #118 — Refactor MainViewModel into Panel-ViewModels (1480 lines too large)
